All the fix for Java 17 support :)
The point of why it can't work on Java 17 I found is mainly because of the configuration system FAWE is using, but because the config system is using reflection stuff and also the structure/usage are really similar to the ConfigLib by Exlll I use. Then, I replaced the whole config system to ConfigLib v4.

Also, as you can see from all the previous commits. Updating all the old stuff is are little frustrating, I also spent a bunch of time for finding those dependencies, some of them are ancient, you can't find too much details about that. Even some dependencies are hosted on my repo lol. There are some "features" are removed, FactionsFeature. Since I can't find the correct version of the api and I dont think anyone in 2022 still using that factions plugin. That why it got removed.

There are one method in ClipboardFormat is removed (It is unused anyway). It because the target method are contains method in ReflectionUtils which is no longer usable in Java 17 (I guess), and idk will it affect plugin compatibility.

BTW, I haven't fully tested the plugin so lol. I didn't make any big changes will affect functionality, so it should be fine :)

-
Some Cool stuff

- Those changes take me like 6 hours to complete lol
- Finding those dependencies is a nightmare
- Making changes for Java 17 support let me lose so many brain cells
- I try my best
- I *almost* cries when I saw the plugin successfully running with Java 17
- Do I need to publish the API?
- **sigh** finally I make it

You can grab the build in release, so you don't need to build it yourself :)